add cbaby ebaby pools (#149)
